Removed max32_alloc_size field from the iova_domain structure.
This field exists for the purpose of optimizing a path of failed
allocations in 32-bit DMA space, when the 32-bit range is depleted.
There are a number of issues:

1. max32_alloc_size is updated in __alloc_and_insert_iova_range()
   without checking the size after allocation of the range fails. This
   will cause a wrong failed small allocation after a "big" allocation
   fails.
2. max32_alloc_size is updated in __cached_rbnode_delete_update()
   without checking the top bound of the released range and the size
   of the released range. This defeats the purpose of the intended
   "optimization".
3. Alignment constaints and DMA  address limits are not taken into account
   when updating this field after the allocation fails. This results in
   future allocation failures that can succeed if looking for available
   ranges is not skipped.

Rather than fixing this "optimization", remove it altogether. We shall
not optimize for failed allocation path that is rare, and the optimizing
affect of this feature is questionable.