Allocate DMA addresses from top to bottom in IOVA domains.
The cached_node and cached32_node fields of the iova_domain structure
are used as the starting point for the address search only if the cached
node starts at the DMA limit or below it, or if the DMA limit is 64 bit
or 32 bit respectively.

The cached_node and cached32_node are updated upon successful allocation
only if the search was performed from the node that does not lie below the
cached values and not above the DMA limit.

For clarity, cached_node field was renamed to cached_top_node.

To enable the existing optimization for network stack behavior - where
network drivers can allocate above 250K DMA buffers for network pools
without using SG tables, we add cached_middle_node and middle_pfn_limit
fields. Without using those, the system locks up for minutes at boot
time trying to allocate network pools.

This ensures contiguous allocations from top to the bottom, with holes
due to alignment or due to lower DMA address requrirements for some devices
in the group. Altogether, this avoids fragmentation of DMA address space
and ensures that large DMA ranges are available.